+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Delete any existing `.vscode_test_results` files at extension activation
 * Changed test results summary displayed using CodeLens in the tests object to re-run only the object tests
 * Add quick fix for the "Missing reference to the built-in protocol" linter warning
+* Add quick fixes for documentation linter warnings
 * Provide limited support for virtual workspaces
 * Fix test results CodeLens showing multiple occurrences of the `(outdated)` text after running individual tests
 * Fix timing issue that could result in multiple Logtalk terminals being created at startup